+++ TestUtilities/Sources/Mocks/DatadogRUM/RUMFeatureMocks.swift
@@ -1709,7 +1709,8 @@ extension AppHang.BacktraceGenerationResult: AnyMockable, RandomMockable {
         return [
             .succeeded(.mockRandom()),
             .failed,
-            .notAvailable
+            .notAvailable,
+            .disabled
         ].randomElement()!
     }
 }